+++ lib/src/extensions/extensions.dart
@@ -1,3 +1,3 @@
 export 'value_listenable_extension.dart';
-export 'value_notifier_extensions.dart';
+export 'value_notifier_extensions/value_notifier_extensions.dart';
 export 'watcher_extensions.dart';